rtsp-recorder/requirements.txt
opencv-python-headless
numpy
schedule
PyYAML>=6.0
python-dotenv
requests